# Add support for foveation during additional render passes