--- a/src/renderers/common/Renderer.js
+++ b/src/renderers/common/Renderer.js
@@ -1419,7 +1419,11 @@ class Renderer {
 
 			}
 
+			const currentXREnabled = this.xr.enabled;
+
+			this.xr.enabled = false;
 			this._renderScene( quad, quad.camera, false );
+			this.xr.enabled = currentXREnabled;
 
 		}
 
@@ -2017,7 +2021,7 @@ class Renderer {
 	 */
 	get currentToneMapping() {
 
-		return this._renderTarget !== null ? NoToneMapping : this.toneMapping;
+		return ( this._renderTarget !== null && this._renderTarget.isXRRenderTarget !== true ) ? NoToneMapping : this.toneMapping;
 
 	}
 
@@ -2029,7 +2033,7 @@ class Renderer {
 	 */
 	get currentColorSpace() {
 
-		return this._renderTarget !== null ? LinearSRGBColorSpace : this.outputColorSpace;
+		return ( this._renderTarget !== null && this._renderTarget.isXRRenderTarget !== true ) ? LinearSRGBColorSpace : this.outputColorSpace;
 
 	}