Proposal to foveate a user provided texture. I need to talk to the immersive group about adding this new API.

Without this feature:
`Surface 2    | 3360x1760 | color 32bit, depth 24bit, stencil 0 bit, MSAA 4, Mode: 1 (HwBinning) | 63  384x256 bins ( 63  rendered) |  1.38 ms | 128 stages : Binning : 0.194ms Render : 0.727ms StoreColor : 0.133ms Blit : 0.004ms`

With this feature:
`Surface 8    | 3360x1760 | color 32bit, depth 24bit, stencil 0 bit, MSAA 4, Mode: 1 (HwBinning) | 72  288x320 bins ( 38  rendered) |  1.26 ms | 78  stages : Binning : 0.161ms Render : 0.514ms StoreColor : 0.31ms Blit : 0.004ms`

Don't pay too much attention to the timings as they depend on the content that is on the screen. The important part is that there are fewer rendered "bins" since they're bigger in the foveated areas.